+++ libs/core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 2.30.0
+
+### Minor Changes
+
+- 678801d: **Button:** Add wrapped boolean to enable wrapping
+
+### Patch Changes
+
+- 678801d: **Button:** Remove the truncating text on button
+- fbaddcb: **Radio:** Fix issue with falsy values
+
+## 2.29.0
+
+### Minor Changes
+
+- 9e72541: **Table:** Add support for search input label
+
+### Patch Changes
+
+- 9e72541: **RBCB:** Change the margin block method
+
 ## 2.28.4
 
 ### Patch Changes